backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
google-generativeai
pydantic
motor